pyproject.toml
[tool.poetry]
name = "fastsocialapi"
version = "0.1.0"
description = ""
authors = ["Mantas"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.95.1"}
psycopg2 = "^2.9.6"
sqlalchemy = "^2.0.12"
bcrypt = "^4.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
